% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_age_comorbidity.R
\name{adjust_age_comorbidity}
\alias{adjust_age_comorbidity}
\title{adjusr_age_comorbidity'}
\usage{
adjust_age_comorbidity(data, comorb, country_select, table = TRUE)
}
\arguments{
\item{data}{- contains id, pre-existing conditions, countries}

\item{comorb}{- preexisting conditions such as : asthma, diabetes_type_one, diabetes_type_two, obesity, hypertension, heart_disease, lung_condition, liver_disease, kidney_disease; 
all pre-existing condition to be passed exactly as written}

\item{country_select}{country which selected. for top 5 countries which has the highest number of responders}

\item{table}{if this TRUE then it will give a long table with all the standard population and study population across age groups. It gives the adjusted comorbidity rate as well in the last column}
}
\value{

}
\description{
adjusr_age_comorbidity'
}
