% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_age_sympt.R
\name{adjusted_age_sympt}
\alias{adjusted_age_sympt}
\title{adjusted_age_sympt}
\usage{
adjusted_age_sympt(data, sympt, country_select, table = TRUE)
}
\arguments{
\item{data}{contains id, age, country,  chills, cough, diarrhoea, fatigue, headache, loss_smell_taste, muscle_ache, nasal_congestion, nausea_vomiting, 
shortness_breath, sore_throat, sputum, temperature, loss_appetite, sneezing, chest_pain, itchy_eyes, joint_pain}

\item{sympt}{contains chills, cough, diarrhoea, fatigue, headache, loss_smell_taste, muscle_ache, nasal_congestion, nausea_vomiting, 
shortness_breath, sore_throat, sputum, temperature, loss_appetite, sneezing, chest_pain, itchy_eyes, joint_pain}

\item{country_select}{which country under analysis}

\item{table}{if TRUE gives a table with all the numbers of standard population and study population}
}
\value{

}
\description{
adjusted_age_sympt
}
