% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_adjusted_comorb_dst.R
\name{age_adjusted_comorb_dst}
\alias{age_adjusted_comorb_dst}
\title{age_adjusted_comorb_dst}
\usage{
age_adjusted_comorb_dst(data, country_select, comorb, table_gen = TRUE)
}
\arguments{
\item{data}{}

\item{country_select}{}

\item{comorb}{}

\item{table}{}
}
\value{

}
\description{
age_adjusted_comorb_dst
}
