% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_adjusted_sympt_dstd.R
\name{age_adjusted_sympt_dst}
\alias{age_adjusted_sympt_dst}
\title{age_adjusted_sympt_dst}
\usage{
age_adjusted_sympt_dst(data, country_select, sympt, table_gen = TRUE)
}
\arguments{
\item{data}{}

\item{country_select}{}

\item{sympt}{}

\item{table}{}
}
\value{

}
\description{
age_adjusted_sympt_dst
}
