% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidities_age_band.R
\name{comorbidities_age_band}
\alias{comorbidities_age_band}
\title{comorbidities_age_band}
\usage{
comorbidities_age_band(
  data,
  start_date = as.Date("2020-04-19"),
  end_date = as.Date("2020-09-01"),
  plot_chart = TRUE
)
}
\arguments{
\item{data}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}
}
\value{

}
\description{
comorbidities_age_band
}
