% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidities_age_band_table.R
\name{comorbidities_age_band_table}
\alias{comorbidities_age_band_table}
\title{comorbidiites_age_band_table}
\usage{
comorbidities_age_band_table(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
comorbidiites_age_band_table
}
