% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidities_covid_status.R
\name{comorbidities_covid_status}
\alias{comorbidities_covid_status}
\title{comorbidities_covid_status}
\usage{
comorbidities_covid_status(
  data,
  start_date = as.Date("2020-04-19"),
  end_date = as.Date("2020-09-01"),
  plot_chart = TRUE
)
}
\arguments{
\item{data}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}
}
\value{

}
\description{
comorbidities_covid_status
}
