% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidities_covsympt.R
\name{comorbidities_covsympt}
\alias{comorbidities_covsympt}
\title{comorbidities_covsympt}
\usage{
comorbidities_covsympt(
  data,
  start_date = as.Date("2020-04-09", tz = "Europe/London"),
  end_date = as.Date("2020-05-06", tz = "Europe/London"),
  plot_chart = TRUE
)
}
\arguments{
\item{data}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}
}
\value{

}
\description{
comorbidities_covsympt
}
