% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity_mild_mod_severe.R
\name{comorbidity_mild_mod_severe}
\alias{comorbidity_mild_mod_severe}
\title{comorbidity_mild_mod_severe}
\usage{
comorbidity_mild_mod_severe(
  data,
  start_date = as.Date("2020-04-09", tz = "Europe/London"),
  end_date = as.Date("2020-05-06", tz = "Europe/London"),
  plot_chart = TRUE,
  breaks_plot = seq(0, 2000, by = 20)
)
}
\arguments{
\item{data}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}
}
\value{

}
\description{
comorbidity_mild_mod_severe
}
