% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_mapp_country.R
\name{count_mapp_country}
\alias{count_mapp_country}
\title{count_map_country}
\usage{
count_mapp_country(data, item = "world", plot_chart = TRUE)
}
\arguments{
\item{data}{}

\item{item}{}

\item{plot_chart}{}
}
\value{

}
\description{
count_map_country
}
