% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_map_world.R
\name{count_mapp_world}
\alias{count_mapp_world}
\title{count_mapp_world}
\usage{
count_mapp_world(
  data,
  item = "world",
  start_date = "2020-04-09",
  end_date = "2020-05-09",
  plot_chart = TRUE,
  title = "World Map"
)
}
\arguments{
\item{data}{}

\item{item}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}

\item{title}{}
}
\value{

}
\description{
count_mapp_world
}
