% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_plot_self_diagnosis.R
\name{count_plot_self_diagnosis}
\alias{count_plot_self_diagnosis}
\title{count_plot_self_diagnosis}
\usage{
count_plot_self_diagnosis(
  data,
  start_date = as.Date("2020-01-01", format = "\%Y-\%m-\%d"),
  end_date = as.Date("2020-01-01", format = "\%Y-\%m-\%d"),
  plot_chart = TRUE,
  title = "Test"
)
}
\arguments{
\item{data}{}

\item{start_date}{}

\item{end_date}{}

\item{title}{}

\item{plot}{}
}
\value{

}
\description{
count_plot_self_diagnosis
}
