% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_map.R
\name{country_map}
\alias{country_map}
\title{country_map}
\usage{
country_map(country, start_date, end_date, plot_chart, which_country)
}
\arguments{
\item{country}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}

\item{which_country}{}
}
\value{

}
\description{
country_map
}
