% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_positive.R
\name{covid_positive}
\alias{covid_positive}
\title{covid_positive}
\usage{
covid_positive(
  data,
  start_date = as.Date("2020-04-09", tz = "Europe/London"),
  end_date = as.Date("2020-05-06", tz = "Europe/London"),
  plot_chart = TRUE
)
}
\arguments{
\item{data}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}
}
\value{

}
\description{
covid_positive
}
