% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_status_symptoms.R
\name{covid_status_symptoms}
\alias{covid_status_symptoms}
\title{covid_status_symptoms}
\usage{
covid_status_symptoms(
  data,
  start_date = as.Date("2020-04-09"),
  end_date = as.Date("2020-09-01"),
  plot_chart = TRUE
)
}
\arguments{
\item{data}{contains id, covid_tested (with categories - positive, negative, showin symptoms) and comorbidities}

\item{start_date}{min date as.Date}

\item{end_date}{max date as.Date}

\item{plot_chart}{if plot_chart == TRUE a plot with percentages of symptoms across covid tested categories}
}
\value{

}
\description{
covid_status_symptoms
}
