% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cstats.table.R
\name{cstats.table}
\alias{cstats.table}
\title{cstats.table}
\usage{
cstats.table(dist, tree, k)
}
\arguments{
\item{dist}{- compute the distance between the variables for clustering}

\item{tree}{- what type of tree to pass (in hierarchical clustering  I have different algoritms - e.g. pass divisive cluster / agglomerative one)}

\item{k}{- number of clusters}
}
\value{

}
\description{
cstats.table
}
