% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diabetes_one_plot.R
\name{diabetes_one_plot}
\alias{diabetes_one_plot}
\title{diabetes_one_plot}
\usage{
diabetes_one_plot(
  data,
  start_date = as.Date("2020-04-09", tz = "Europe/London"),
  end_date = as.Date("2020-08-20"),
  plot_chart = TRUE
)
}
\arguments{
\item{data}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}
}
\value{

}
\description{
diabetes_one_plot
}
