% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_age_band.R
\name{gender_age_band}
\alias{gender_age_band}
\title{gender_age_band}
\usage{
gender_age_band(data, age_category = "20-39", plot_chart = TRUE)
}
\arguments{
\item{data}{contains id, age_band and gender}

\item{age_category}{categories to choose are 20-39; 40-59; 60+}

\item{plot_chart}{if plot_chart == TRUE gives as an output a pie chart of percentages of gender, if FALSE it gives me a table with counts and percentages of gender accross age band chosen}
}
\value{

}
\description{
gender_age_band
}
