% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_covid_status.R
\name{gender_covid_status}
\alias{gender_covid_status}
\title{gender_covid_status}
\usage{
gender_covid_status(data, covid_status = "negative", plot_chart = TRUE)
}
\arguments{
\item{data}{}

\item{covid_status}{}

\item{plot_chart}{}
}
\value{

}
\description{
gender_covid_status
}
