% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_distrib.R
\name{gender_distrib}
\alias{gender_distrib}
\title{gender_distrib}
\usage{
gender_distrib(
  data,
  start_date = as.Date("2020-04-09", tz = "Europe/London"),
  end_date = as.Date("2020-05-06", tz = "Europe/London"),
  plot_chart = TRUE
)
}
\arguments{
\item{data}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}
}
\value{

}
\description{
gender_distrib
}
