% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertension_plot.R
\name{hypertension_plot}
\alias{hypertension_plot}
\title{hypertension_num}
\usage{
hypertension_plot(
  data = data,
  start_date = as.Date("2020-04-09", tz = "Europe/London"),
  end_date = as.Date("2020-08-20"),
  plot_chart = TRUE
)
}
\arguments{
\item{data}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}
}
\value{

}
\description{
hypertension_num
}
