% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lung_disease_covsympt.R
\name{lung_disease_covsympt}
\alias{lung_disease_covsympt}
\title{lung_disease_covsympt}
\usage{
lung_disease_covsympt(
  data,
  start_date = as.Date("2020-04-09", tz = "Europe/London"),
  end_date = as.Date("2020-05-06", tz = "Europe/London"),
  plot_chart = TRUE
)
}
\arguments{
\item{data}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}
}
\value{

}
\description{
lung_disease_covsympt
}
