% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lung_plot.R
\name{lung_plot}
\alias{lung_plot}
\title{lung_plot}
\usage{
lung_plot(
  data,
  start_date = as.Date("2020-04-09", tz = "Europe/London"),
  end_date = as.Date("2020-08-20"),
  plot_chart = TRUE
)
}
\arguments{
\item{data}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}
}
\value{

}
\description{
lung_plot
}
