% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_severe_prof.R
\name{map_severe_prof}
\alias{map_severe_prof}
\title{map_severe_prof}
\usage{
map_severe_prof(
  data,
  item = "world",
  start_date = as.Date("2020-04-01", format = "\%Y-\%m-\%d"),
  end_date = as.Date("2020-05-09", format = "\%Y-\%m-\%d"),
  plot_chart = TRUE
)
}
\arguments{
\item{data}{}

\item{item}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}
}
\value{

}
\description{
map_severe_prof
}
