% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mild_comorbid_symptoms.R
\name{mild_comorbid_symptoms}
\alias{mild_comorbid_symptoms}
\title{mild_comorbid_symptoms}
\usage{
mild_comorbid_symptoms(
  data,
  start_date = as.Date("2020-04-09", tz = "Europe/London"),
  end_date = as.Date("2020-05-06", tz = "Europe/London"),
  plot_chart = TRUE,
  breaks_plot = seq(0, 2000, by = 100)
)
}
\arguments{
\item{data}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}

\item{breaks_plot}{}
}
\value{

}
\description{
mild_comorbid_symptoms
}
