% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderate_manifestation_plot.R
\name{moderate_manifestation_plot}
\alias{moderate_manifestation_plot}
\title{moderate_manifestation_plot}
\usage{
moderate_manifestation_plot(
  data,
  start_date = as.Date("2020-04-09", format = "\%Y-\%m-\%d"),
  end_date = as.Date("2020-05-09", format = "\%Y-\%m-\%d"),
  plot_chart = TRUE
)
}
\arguments{
\item{data}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}
}
\value{

}
\description{
moderate_manifestation_plot
}
