% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_cov.R
\name{no_cov}
\alias{no_cov}
\title{no_cov}
\usage{
no_cov(
  data,
  start_date = as.Date("2020-04-09", tz = "Europe/London"),
  end_date = as.Date("2020-05-06", tz = "Europe/London"),
  plot_chart = TRUE
)
}
\arguments{
\item{data}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}
}
\value{

}
\description{
no_cov
}
