% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_covid_symptoms.R
\name{no_covid_symptoms}
\alias{no_covid_symptoms}
\title{no_covid_symptoms}
\usage{
no_covid_symptoms(
  data,
  start_date = as.Date("2020-04-09", tz = "Europe/London"),
  end_date = as.Date("2020-05-06", tz = "Europe/London"),
  plot_chart = TRUE
)
}
\arguments{
\item{data}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}
}
\value{

}
\description{
no_covid_symptoms
}
