% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obesity_covsympt.R
\name{obesity_covsympt}
\alias{obesity_covsympt}
\title{obesity_covsympt}
\usage{
obesity_covsympt(
  data,
  start_date = as.Date("2020-04-09", tz = "Europe/London"),
  end_date = as.Date("2020-05-06", tz = "Europe/London"),
  plot_chart = TRUE
)
}
\arguments{
\item{data}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}
}
\value{

}
\description{
obesity_covsympt
}
