% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obesity_plot.R
\name{obesity_plot}
\alias{obesity_plot}
\title{obesity_plot}
\usage{
obesity_plot(
  data = data,
  start_date = as.Date("2020-04-09", tz = "Europe/London"),
  end_date = as.Date("2020-08-20"),
  plot_chart = TRUE
)
}
\arguments{
\item{data}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}
}
\value{

}
\description{
obesity_plot
}
