% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_adjusted_preexisting_rates.R
\name{plot_adjusted_preexisting_rates}
\alias{plot_adjusted_preexisting_rates}
\title{plot_adjusted_preexisting_rates}
\usage{
plot_adjusted_preexisting_rates(data, title = "Exampele", plot_chart = TRUE)
}
\arguments{
\item{data}{}

\item{title}{}

\item{plot_chart}{}
}
\value{

}
\description{
plot_adjusted_preexisting_rates
}
