% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_plot_self_diagnosis.R
\name{self_diagnosis_mild_mod_severe}
\alias{self_diagnosis_mild_mod_severe}
\title{self_diagnosis_mild_mod_severe}
\usage{
self_diagnosis_mild_mod_severe(
  data,
  start_date = as.Date("2020-01-01", format = "\%Y-\%m-\%d"),
  end_date = as.Date("2020-01-01", format = "\%Y-\%m-\%d"),
  plot_chart = TRUE,
  title = "Test"
)
}
\arguments{
\item{data}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}

\item{title}{}
}
\value{

}
\description{
self_diagnosis_mild_mod_severe
}
