% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severe_comorbid_symptoms.R
\name{severe_comorbid_symptoms}
\alias{severe_comorbid_symptoms}
\title{severe_comorbid_symptoms}
\usage{
severe_comorbid_symptoms(
  data,
  start_date = as.Date("2020-04-09", tz = "Europe/London"),
  end_date = as.Date("2020-05-06", tz = "Europe/London"),
  plot_chart = TRUE,
  breaks_plot = seq(0, 100, by = 20)
)
}
\arguments{
\item{data}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}

\item{breaks_plot}{}
}
\value{

}
\description{
severe_comorbid_symptoms
}
