% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_symptoms_cov.R
\name{show_symptoms_cov}
\alias{show_symptoms_cov}
\title{show_symptoms_cov}
\usage{
show_symptoms_cov(
  data,
  start_date = as.Date("2020-04-09", tz = "Europe/London"),
  end_date = as.Date("2020-05-06", tz = "Europe/London"),
  plot_chart = TRUE
)
}
\arguments{
\item{data}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}
}
\value{

}
\description{
show_symptoms_cov
}
