% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sympt_negative_age_band.R
\name{sympt_negative_age_band}
\alias{sympt_negative_age_band}
\title{sympt_negative_age_band}
\usage{
sympt_negative_age_band(
  data,
  start_date = as.Date("2020-04-19"),
  end_date = as.Date("2020-09-01"),
  plot_chart = TRUE
)
}
\arguments{
\item{data}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}
}
\value{

}
\description{
sympt_negative_age_band
}
