% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sympt_positive_age_band.R
\name{sympt_positive_age_band}
\alias{sympt_positive_age_band}
\title{sympt_positive_age_band}
\usage{
sympt_positive_age_band(
  data,
  start_date = as.Date("2020-04-19"),
  end_date = as.Date("2020-09-01"),
  plot_chart = TRUE
)
}
\arguments{
\item{data}{a dataset that contains ids, showing symptoms of covid, and symptoms}

\item{start_date}{a min date passed in as.Date}

\item{end_date}{a max data passed as.Date}

\item{plot_chart}{if TRUE then give me a chart, if FALSE then give me a table with symptoms accross age, with counts and percentages}
}
\value{

}
\description{
sympt_positive_age_band
}
