% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symptom_comorbidities_no_cv.R
\name{symptom_comorbidities_no_cv}
\alias{symptom_comorbidities_no_cv}
\title{symptom_comorbidities_no_cv counting non covid}
\usage{
symptom_comorbidities_no_cv(
  data = data,
  start_date = as.Date("2020-04-09", tz = "Europe/London"),
  end_date = as.Date("2020-05-06", tz = "Europe/London"),
  plot_chart = TRUE
)
}
\arguments{
\item{data}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}
}
\value{

}
\description{
symptom_comorbidities_no_cv counting non covid
}
