% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symptom_profile_covid_tested.R
\name{symptom_profile_covid_tested}
\alias{symptom_profile_covid_tested}
\title{symptom_profile_covid_tested}
\usage{
symptom_profile_covid_tested(
  data,
  start_date = as.Date("2020-01-01", format = "\%Y-\%m-\%d"),
  end_date = as.Date("2020-02-01", format = "\%Y-\%m-\%d"),
  plot_chart = TRUE
)
}
\arguments{
\item{data}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}

\item{title}{}
}
\value{

}
\description{
symptom_profile_covid_tested
}
