% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symptom_profile_frequency_plot.R
\name{symptom_profile_frequency_plot}
\alias{symptom_profile_frequency_plot}
\title{symptom_profile_frequency_plot}
\usage{
symptom_profile_frequency_plot(
  data,
  start_date = as.Date("2020-04-09", format = "\%Y-\%m-\%d"),
  end_date = as.Date("2020-05-09", format = "\%Y-\%m-\%d"),
  plot_chart = TRUE,
  title = "Testt"
)
}
\arguments{
\item{data}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}

\item{title}{}
}
\value{

}
\description{
symptom_profile_frequency_plot
}
