% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symptom_profile_showing_symptom.R
\name{symptom_profile_showing_symptom}
\alias{symptom_profile_showing_symptom}
\title{symptom_profile_showing_symptom}
\usage{
symptom_profile_showing_symptom(
  data,
  start_date = as.Date("2020-04-09", format = "\%Y-\%m-\%d"),
  end_date = as.Date("2020-05-09", format = "\%Y-\%m-\%d"),
  plot_chart = TRUE
)
}
\arguments{
\item{data}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}
}
\value{

}
\description{
symptom_profile_showing_symptom
}
