% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symptom_profile_test.R
\name{symptom_profile_test}
\alias{symptom_profile_test}
\title{system_profile_test}
\usage{
symptom_profile_test(
  data,
  start_date = as.Date("2020-01-01", format = "\%Y-\%m-\%d"),
  end_date = as.Date("2020-02-01", format = "\%Y-\%m-\%d"),
  plot_chart = TRUE,
  title = "Testt"
)
}
\arguments{
\item{data}{}

\item{start_date}{}

\item{end_date}{}

\item{plot_chart}{}

\item{title}{}
}
\value{

}
\description{
system_profile_test
}
