% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_for_data.R
\name{count_plot_self_diagnosis}
\alias{count_plot_self_diagnosis}
\title{count_plot_self_diagnosis}
\usage{
count_plot_self_diagnosis(
  data,
  start_date,
  end_date,
  plot = TRUE,
  title = "Test"
)
}
\arguments{
\item{data}{}

\item{start_date}{}

\item{end_date}{}

\item{plot}{}

\item{title}{}
}
\value{

}
\description{
count_plot_self_diagnosis
}
\examples{
data 
}
